--- lib/ddsketch/version.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module DDSketch
-  module VERSION
-    MAJOR = 0
-    MINOR = 1
-    PATCH = 0
-    PRE = nil
-
-    STRING = [MAJOR, MINOR, PATCH, PRE].compact.join('.')
-  end
-end
