+++ spec/datadog/ddsketch/ddsketch_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+describe Datadog::DDSketch::Sketch do
+  it_behaves_like 'test sketch', { relative_accuracy: 0.05 }
+end
