+++ spec/datadog/ddsketch/log_collapsing_lowest_dense_sketch_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+describe Datadog::DDSketch::LogCollapsingLowestDenseSketch do
+  it_behaves_like 'test sketch', { relative_accuracy: 0.05, bin_limit: 1024 }
+end
