+++ spec/datadog/ddsketch/logarithmic_key_mapping_spec.rb
@@ -1,0 +1,5 @@
+describe Datadog::DDSketch::LogarithmicKeyMapping do
+  include_context 'mapping tests' do
+    let(:mapping) { described_class.new(relative_accuracy: relative_accuracy, offset: offset) }
+  end
+end
