+++ spec/spec_helper.rb
@@ -24,9 +24,10 @@ end
 # it.
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
-
+require 'pry'
 Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
-require 'ddsketch'
+
+require 'datadog/ddsketch'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
